- renamed node.js scripts for clarity - pulling in env variables including optimize flag
